Added Delete Address Endpoint, also changed Rules

+++ src/main/java/com/ahmetaksunger/ecommerce/service/PaymentDetailManager.java
@@ -39,7 +39,7 @@ public class PaymentDetailManager implements PaymentDetailService{
     @Override
     public void delete(long paymentDetailId,User user) {
         // Rules
-        paymentDetailRules.canDeleteAccount(paymentDetailId,user);
+        paymentDetailRules.checkIfCanDelete(paymentDetailId,user);
 
         paymentDetailRepository.deleteById(paymentDetailId);
     }
